+++ CleverPush/Source/CPAppBannerModule.h
@@ -30,6 +30,7 @@
 + (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId;
 + (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId force:(BOOL)force;
 + (void)setBannerOpenedCallback:(CPAppBannerActionBlock)callback;
++ (void)setBannerShownCallback:(CPAppBannerShownBlock)callback;
 + (void)getBanners:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId groupId:(NSString*)groupId completion:(void(^)(NSMutableArray<CPAppBanner*>*))callback;
 + (void)presentAppBanner:(UIViewController*)controller banner:(CPAppBanner*)banner;
 + (void)showNextActivePendingBanner:(CPAppBanner*)banner;