+++ CleverPush/Source/CleverPushInstance.m
@@ -71,7 +71,7 @@
 
 @implementation CleverPushInstance
 
-NSString * const CLEVERPUSH_SDK_VERSION = @"1.28.4";
+NSString * const CLEVERPUSH_SDK_VERSION = @"1.28.7";
 
 static BOOL registeredWithApple = NO;
 static BOOL startFromNotification = NO;
@@ -3171,6 +3171,10 @@ static id isNil(id object) {
 
 - (void)setAppBannerOpenedCallback:(CPAppBannerActionBlock)callback {
     [CPAppBannerModule setBannerOpenedCallback:callback];
+}
+
+- (void)setAppBannerShownCallback:(CPAppBannerShownBlock)callback {
+    [CPAppBannerModule setBannerShownCallback:callback];
 }
 
 - (void)disableAppBanners {